+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Fix crash when remote player enters after a pvp battle is done
 - Fix ig.ENTITY.OneTimeSwitch not switching to off animation on remote
+- Fix SHOW_EXTERN_ANIM anims syncing only for the current player on remote
 
 ## [0.7.5] 2026-03-06